split code for browsing
